+++ src/test/java/com/padmaja/objectcomparison/TestTest.java
@@ -1,7 +1,6 @@
-package com.padmaja.stringreverse;
+package com.padmaja.objectcomparison;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.jupiter.api.Assertions.*;
 
 class TestTest {
 
@@ -25,7 +24,6 @@ class TestTest {
         System.out.println(b2.equals(b3));
         boolean expResults = true;
         assertEquals(expResults, Result2);
-
 
         boolean Result3 = b1.equals(new String("abc"));
         System.out.println(Result3);
